pyproject.toml
[tool.poetry]
name = "datalake"
version = "0.1"
description = "assignment demo"
authors = ["Dinesh Ram D"]

packages = [
    { include = "silver_layer", from = "datalake" },
    { include = "gold_layer", from = "datalake" },
    { include = "bronze_layer", from = "datalake"},
  {include = "datalake"}
]
license = "-"
#readme = "README.md"
[tool.poetry.group.main.dependencies]
sf-hamilton = {extras = ["visualization"], version = "^1.86.1"}
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
confluent-kafka = "^2.3.0"
kafka-python = {git = "https://github.com/dpkp/kafka-python.git"}
kafka-python-ng = {git = "https://github.com/wbarnha/kafka-python-ng.git"}

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
aiohappyeyeballs = "2.4.4"
aiohttp = "3.11.11"
aiosignal = "1.3.2"
aiosqlite = "0.20.0"
alembic = "1.14.1"
annotated-types = "0.7.0"
anyio = "4.8.0"
apache-airflow = {extras = ["kafka"], version = "^2.10.4"}
apache-airflow-providers-common-compat = "1.3.0"
apache-airflow-providers-common-io = "1.5.0"
apache-airflow-providers-common-sql = "1.21.0"
apache-airflow-providers-fab = "1.5.2"
apache-airflow-providers-ftp = "3.12.0"
apache-airflow-providers-http = "5.0.0"
apache-airflow-providers-imap = "3.8.0"
apache-airflow-providers-smtp = "1.9.0"
apache-airflow-providers-sqlite = "4.0.0"
apispec = "6.8.1"
argcomplete = "3.5.3"
asgiref = "3.8.1"
asttokens = "3.0.0"
attrs = "24.3.0"
babel = "2.16.0"
blinker = "1.9.0"
cachelib = "0.9.0"
certifi = "2024.12.14"
cffi = "1.17.1"
charset-normalizer = "3.4.1"
click = "8.1.8"
clickclick = "20.10.2"
colorama = "0.4.6"
colorlog = "6.9.0"
comm = "0.2.2"
configupdater = "3.2"
connexion = "2.14.2"
cron-descriptor = "1.4.5"
croniter = "6.0.0"
cryptography = "44.0.0"
debugpy = "1.8.12"
decorator = "5.1.1"
deprecated = "1.2.15"
dill = "0.3.9"
dnspython = "2.7.0"
duckdb = "1.1.3"
email-validator = "2.2.0"
executing = "2.2.0"
fastjsonschema = "2.21.1"
ffmpeg = "1.4"
flask = "2.2.5"
flask-appbuilder = "4.5.2"
flask-babel = "2.0.0"
flask-caching = "2.3.0"
flask-jwt-extended = "4.7.1"
flask-limiter = "3.10.1"
flask-login = "0.6.3"
flask-session = "0.5.0"
flask-sqlalchemy = "2.5.1"
flask-wtf = "1.2.2"
frozenlist = "1.5.0"
fsspec = "2024.12.0"
google-re2 = "1.1.20240702"
googleapis-common-protos = "1.66.0"
greenlet = "3.1.1"
grpcio = "1.69.0"
gunicorn = "23.0.0"
h11 = "0.14.0"
httpcore = "1.0.7"
httpx = "0.28.1"
idna = "3.10"
importlib-metadata = "8.5.0"
inflection = "0.5.1"
iniconfig = "2.0.0"
ipykernel = "6.29.5"
ipython = "8.31.0"
itsdangerous = "2.2.0"
jedi = "0.19.2"
jinja2 = "3.1.5"
jmespath = "1.0.1"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
jupyter-client = "8.6.3"
jupyter-core = "5.7.2"
jupytext = "1.16.6"
kafka-python = {git = "https://github.com/dpkp/kafka-python.git"}
lazy-object-proxy = "1.10.0"
limits = "4.0.1"
linkify-it-py = "2.0.3"
lockfile = "0.12.2"
mako = "1.3.8"
markdown-it-py = "3.0.0"
markupsafe = "3.0.2"
marshmallow = "3.26.0"
marshmallow-oneofschema = "3.1.1"
marshmallow-sqlalchemy = "0.28.2"
matplotlib-inline = "0.1.7"
mdit-py-plugins = "0.4.2"
mdurl = "0.1.2"
methodtools = "0.4.7"
more-itertools = "10.6.0"
multidict = "6.1.0"
mypy-extensions = "1.0.0"
nbformat = "5.10.4"
nest-asyncio = "1.6.0"
numpy = "2.2.2"
opentelemetry-api = "1.29.0"
opentelemetry-exporter-otlp = "1.29.0"
opentelemetry-exporter-otlp-proto-common = "1.29.0"
opentelemetry-exporter-otlp-proto-grpc = "1.29.0"
opentelemetry-exporter-otlp-proto-http = "1.29.0"
opentelemetry-proto = "1.29.0"
opentelemetry-sdk = "1.29.0"
opentelemetry-semantic-conventions = "0.50b0"
ordered-set = "4.1.0"
packaging = "24.2"
pandas = "2.2.3"
parso = "0.8.4"
pathspec = "0.12.1"
pendulum = "3.0.0"
platformdirs = "4.3.6"
pluggy = "1.5.0"
prison = "0.2.1"
prompt-toolkit = "3.0.50"
propcache = "0.2.1"
protobuf = "5.29.3"
psutil = "6.1.1"
pure-eval = "0.2.3"
pycparser = "2.22"
pydantic = "2.10.5"
pydantic-core = "2.27.2"
pygments = "2.19.1"
pyjwt = "2.10.1"
pytest = "8.3.4"
python-daemon = "3.1.2"
python-dateutil = "2.9.0.post0"
python-nvd3 = "0.16.0"
python-slugify = "8.0.4"
pytz = "2024.2"
pywin32 = "308"
pyyaml = "6.0.2"
pyzmq = "26.2.0"
referencing = "0.36.1"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
rfc3339-validator = "0.1.4"
rich = "13.9.4"
rich-argparse = "1.6.0"
rpds-py = "0.22.3"
setproctitle = "1.3.4"
sf-hamilton = {extras = ["visualization"], version = "^1.86.1"}
six = "1.17.0"
sniffio = "1.3.1"
sqlalchemy = "1.4.54"
sqlalchemy-jsonfield = "1.0.2"
sqlalchemy-utils = "0.41.2"
sqlparse = "0.5.3"
stack-data = "0.6.3"
tabulate = "0.9.0"
tenacity = "9.0.0"
termcolor = "2.5.0"
text-unidecode = "1.3"
time-machine = "2.16.0"
tornado = "6.4.2"
traitlets = "5.14.3"
typing-inspect = "0.9.0"
typing-extensions = "4.12.2"
tzdata = "2025.1"
uc-micro-py = "1.0.3"
universal-pathlib = "0.2.6"
urllib3 = "2.3.0"
watchdog = "6.0.0"
wcwidth = "0.2.13"
werkzeug = "2.2.3"
wirerope = "1.0.0"
wrapt = "1.17.2"
wtforms = "3.2.1"
yarl = "1.18.3"
zipp = "3.21.0"
python-dotenv = "^1.0.1"
kafka-python-ng = {git = "https://github.com/wbarnha/kafka-python-ng.git"}
ruff = "^0.9.3"
graphviz = "^0.20.3"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
